+++ data-raw/create sysdata.R
@@ -6,18 +6,21 @@ logo <- magick::image_read_pdf("data-raw/GrattanSVGLogo.pdf")
 logogrob <- grid::rasterGrob(logo)
 
 chart_types <- tibble::tribble(
-                         ~type, ~width, ~height, ~caption, ~title,  ~subtitle, ~class, ~pptx_template,
-                      "normal",  22.16,    14.5,      120,     70,         75, "normal", "template_normal.pptx",
-                  "normal_169",  30.00,    14.5,      180,     95,        100, "normal", NA_character_,
-                        "tiny",  22.16,   11.08,      120,     70,         75, "normal", NA_character_,
-                 "wholecolumn",  22.16,   22.16,      120,     70,         75, "normal", "template_wholecolumn.pptx",
-                    "fullpage",  44.32,   22.16,      240,    140,        150, "normal", "template_fullpage.pptx",
-                   "fullslide",   25.4,   19.05,      140,     55,         70, "fullslide", "template_43.pptx",
-               "fullslide_169",   33.87, 19.05,       175,     55,         95, "fullslide", "template_169.pptx",
-            "fullslide_old169",   25.4,  14.29,       140,     55,         70, "fullslide", "template_old_169.pptx",
-                "fullslide_44",   25.4,  25.4,        140,     55,         95, "fullslide", NA_character_,
-                        "blog",   25.4,   19.05,      155,     62,         85, "fullslide", "template_blog.pptx",
-                   "blog_half", 25.4/2,  19.05,       155,     62,         85, "fullslide", "template_blog_half.pptx")
+                         ~type, ~status,      ~width, ~height, ~caption, ~title,  ~subtitle, ~class,     ~pptx_template,
+                      "normal", "active",        22.16,   14.50,    120,     70,         75, "normal",    "template_normal.pptx",
+                 "wholecolumn", "active",        22.16,   22.16,    120,     70,         75, "normal",    "template_wholecolumn.pptx",
+                    "fullpage", "active",        44.32,   22.16,    240,    140,        150, "normal",    "template_fullpage.pptx",
+                  "normal_169", "active",        30.00,   14.50,    180,     95,        100, "normal",    NA_character_,
+                        "tiny", "active",        22.16,   11.08,    120,     70,         75, "normal",    NA_character_,
+
+                   "fullslide", "active",        33.87,   19.05,    175,     55,         95, "fullslide", "template_169.pptx",
+                        "blog", "active",        25.40,   19.05,    155,     62,         85, "fullslide", "template_blog.pptx",
+
+                "fullslide_44", "depreciated",   25.40,   25.40,    140,     55,         95, "fullslide", NA_character_,
+                "fullslide_43", "depreciated",   25.40,   19.05,    140,     55,         70, "fullslide", "template_43.pptx",
+                   "blog_half", "depreciated",  25.4/2,   19.05,    155,     62,         85, "fullslide", "template_blog_half.pptx",
+            "fullslide_old169", "depreciated",   25.40,   14.29,    140,     55,         70, "fullslide", "template_old_169.pptx",
+)
 
 
 blog_border <- 0.15
@@ -29,25 +32,33 @@ chart_types <- chart_types %>%
                                 TRUE ~ 0.7),
          bottom_border = case_when(class == "normal" ~ 0,
                                    grepl("blog", type) ~ 0.05,
-                                   type %in% c("fullslide_169",
+                                   type %in% c("fullslide",
                                                "fullslide_44") ~ 0.24,
                                    TRUE ~ 0.05),
          left_border = case_when(class == "normal" ~ 0,
-                                 type %in% c("fullslide",
+                                 type %in% c("fullslide_43",
                                              "fullslide_44") ~ (width - 22.16) / 2,
-                                 type == "fullslide_169" ~ (width - 30) / 2,
+                                 type == "fullslide" ~ (width - 30) / 2,
                                  type == "fullslide_old169" ~ (width - 22.64) / 2,
                                  grepl("blog", type) ~ blog_border),
          right_border = left_border)
 
-all_chart_types <- chart_types$type
+chart_types_inc_depreciated <- chart_types
+chart_types <- chart_types[chart_types$status == "active", ]
 
-fullslide_chart_types <- chart_types$type[chart_types$class == "fullslide"]
+all_chart_types <- chart_types$type[chart_types$status == "active"]
+all_chart_types_inc_depreciated <- chart_types$type
+
+fullslide_chart_types <- chart_types$type[chart_types$class == "fullslide" & chart_types$status == "active"]
+fullslide_chart_types_inc_depreciated <- chart_types$type[chart_types$class == "fullslide"]
 
 use_data(logogrob,
          chart_types,
+         chart_types_inc_depreciated,
          all_chart_types,
+         all_chart_types_inc_depreciated,
          fullslide_chart_types,
+         fullslide_chart_types_inc_depreciated,
          internal = TRUE,
          overwrite = TRUE)