+++ tests/testthat/test-grattan_save.R
@@ -32,12 +32,10 @@ test_that("grattan_save() saves charts (no powerpoint)", {
 
   grattan_save(filename = "../figs/grattan_save/test_plot.png",
                object = test_plot,
-               save_pptx = F,
-               save_data = F,
+               save_pptx = FALSE,
+               save_data = FALSE,
                type = "all")
 
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_44.png"))
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_169.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullpage.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_normal_169.png"))
@@ -45,6 +43,9 @@ test_that("grattan_save() saves charts (no powerpoint)", {
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_tiny.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_wholecolumn.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_blog.png"))
+
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_43.png"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_44.png"))
 
   unlink("../figs/grattan_save", recursive = TRUE)
   unlink("../testthat/Rplots.pdf")
@@ -60,8 +61,6 @@ test_that("grattan_save() saves charts (with powerpoint)", {
                save_data = TRUE,
                type = "all")
 
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_44.png"))
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_169.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullpage.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_normal_169.png"))
@@ -71,7 +70,11 @@ test_that("grattan_save() saves charts (with powerpoint)", {
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_blog.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot.xlsx"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide.pptx"))
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_169.pptx"))
+  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullpage.pptx"))
+
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_43.png"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_blog_half.png"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_44.png"))
 
   unlink("../figs/grattan_save", recursive = TRUE)
   unlink("../testthat/Rplots.pdf")
@@ -87,7 +90,7 @@ test_that("grattan_save() saves last_plot() and works with repeated calls", {
   # Test that a repeated save causes no problems
   grattan_save("../figs/grattan_save/test.png", type = "all", save_pptx = TRUE)
 
-  expect_true(file.exists("../figs/grattan_save/test/test_fullslide_44.png"))
+  expect_true(file.exists("../figs/grattan_save/test/test_fullslide.png"))
 
   unlink("../figs/grattan_save", recursive = TRUE)
   unlink("../testthat/Rplots.pdf")
@@ -104,8 +107,6 @@ test_that("grattan_save() doesn't save chart data / PPTX when not requested", {
                object = test_plot,
                type = "all")
 
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_44.png"))
-  expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_169.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_fullpage.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_normal_169.png"))
@@ -113,9 +114,14 @@ test_that("grattan_save() doesn't save chart data / PPTX when not requested", {
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_tiny.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_wholecolumn.png"))
   expect_true(file.exists("../figs/grattan_save/test_plot/test_plot_blog.png"))
-  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot.xlsx"))
+
   expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide.pptx"))
-  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_169.pptx"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_blog.pptx"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot.xlsx"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_43.png"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_blog_half.png"))
+  expect_false(file.exists("../figs/grattan_save/test_plot/test_plot_fullslide_44.png"))
+
 
   unlink("../figs/grattan_save", recursive = TRUE)
   unlink("../testthat/Rplots.pdf")
@@ -128,10 +134,10 @@ test_that("grattan_save() saves chart data when requested",{
                type = "normal",
                save_data = TRUE)
 
-  expect_true(file.exists("test.pdf"))
-  expect_true(file.exists("test.xlsx"))
+  expect_true(file.exists("test/test_normal.pdf"))
+  expect_true(file.exists("test/test.xlsx"))
 
-  saved_data <- openxlsx::read.xlsx("test.xlsx",
+  saved_data <- openxlsx::read.xlsx("test/test.xlsx",
                                     rows = c(3:35),
                                     cols = c(2:12))
 
@@ -159,36 +165,36 @@ test_that("grattan_save() height behaviour works as expected with normal charts"
                type = "normal",
                height = 20)
 
-  expect_true(file.exists("test_plot_normal_default_height.png"))
-  expect_true(file.exists("test_plot_normal_manual_height.png"))
+  expect_true(file.exists("test_plot_normal_default_height/test_plot_normal_default_height_normal.png"))
+  expect_true(file.exists("test_plot_normal_manual_height/test_plot_normal_manual_height_normal.png"))
 
-  expect_gt(file.size("test_plot_normal_manual_height.png"),
-            file.size("test_plot_normal_default_height.png"))
+  expect_gt(file.size("test_plot_normal_manual_height/test_plot_normal_manual_height_normal.png"),
+            file.size("test_plot_normal_default_height/test_plot_normal_default_height_normal.png"))
 
-  unlink("test_plot_normal_default_height.png")
-  unlink("test_plot_normal_manual_height.png")
+  unlink("test_plot_normal_default_height", recursive = TRUE)
+  unlink("test_plot_normal_manual_height", recursive = TRUE)
 
 })
 
 test_that("grattan_save() height behaviour works as expected with fullslide charts", {
 
-  grattan_save(filename = "test_plot_fullslide_default_height.png",
+  grattan_save(filename = "default_height.png",
                object = test_plot,
                type = "fullslide")
 
-  grattan_save(filename = "test_plot_fullslide_manual_height.png",
+  grattan_save(filename = "manual_height.png",
                object = test_plot,
                type = "fullslide",
                height = 40)
 
-  expect_true(file.exists("test_plot_fullslide_default_height.png"))
-  expect_true(file.exists("test_plot_fullslide_manual_height.png"))
+  expect_true(file.exists("default_height/default_height_fullslide.png"))
+  expect_true(file.exists("manual_height/manual_height_fullslide.png"))
 
-  expect_gt(file.size("test_plot_fullslide_manual_height.png"),
-            file.size("test_plot_fullslide_default_height.png"))
+  expect_gt(file.size("manual_height/manual_height_fullslide.png"),
+            file.size("default_height/default_height_fullslide.png"))
 
-  unlink("test_plot_fullslide_default_height.png")
-  unlink("test_plot_fullslide_manual_height.png")
+  unlink("default_height", recursive = TRUE)
+  unlink("manual_height", recursive = TRUE)
 
 })
 
@@ -208,10 +214,10 @@ test_that("grattan_save() saves a plot with a watermark", {
                object = test_plot,
                watermark = "DRAFT")
 
-  expect_true(file.exists("test_plot_watermark.png"))
+  expect_true(file.exists("test_plot_watermark/test_plot_watermark_normal.png"))
 
 
-  unlink("test_plot_watermark.png")
+  unlink("test_plot_watermark", recursive = TRUE)
 })
 
 
@@ -222,8 +228,6 @@ test_that("grattan_save_all() works", {
   grattan_save_all(filename = "../figs/grattan_save_all/test_plot.png",
                    object = test_plot)
 
-  expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide_44.png"))
-  expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide_169.png"))
   expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide.png"))
   expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullpage.png"))
   expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_normal_169.png"))
@@ -233,7 +237,11 @@ test_that("grattan_save_all() works", {
   expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_blog.png"))
   expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot.xlsx"))
   expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide.pptx"))
-  expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide_169.pptx"))
+  expect_true(file.exists("../figs/grattan_save_all/test_plot/test_plot_blog.pptx"))
+
+  expect_false(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide_43.png"))
+  expect_false(file.exists("../figs/grattan_save_all/test_plot/test_plot_blog_half.png"))
+  expect_false(file.exists("../figs/grattan_save_all/test_plot/test_plot_fullslide_44.png"))
 
   unlink("../figs/grattan_save_all", recursive = TRUE)
   unlink("../testthat/Rplots.pdf")
@@ -243,12 +251,12 @@ test_that("grattan_save_all() works", {
 test_that("grattan_save(ignore_long_titles = TRUE) successfully ignores long titles", {
 
   expect_error(
-    grattan_save(filename = "test_plot_fullslide_default_height.png",
+    grattan_save(filename = "default_height.png",
                object = test_plot_longlabs,
                type = "fullslide")
   )
 
-  grattan_save(filename = "test_plot_fullslide_default_height.png",
+  grattan_save(filename = "default_height.png",
                object = test_plot_longlabs,
                type = "fullslide",
                ignore_long_title = TRUE)