+++ NEWS.md
@@ -1,6 +1,11 @@
+# grattantheme 0.9.0
+* blue added to palette
+* 8 tint (fade) options added to palette
+* depreciated some export formats
+
 # grattantheme 0.8.4
 * path behaviour of export_latex_code() changed
-* 
+
 
 # grattantheme 0.8.3.900
 * When your chart has no subtitle, the chart will not leave space for it